Split global and normal logger tests (#15700)
Move logp global functions to global.go and selective.go, plus split up
tests. The change also adds the `nologpglobal` build tag to global.go and global_test.go.
By default all global functions are available. When compiling or
running a package its test code with `-tags=nologpglobal`, then global
logp functions are not available.

(cherry picked from commit 4183ba365a4c3fe9b8f7ca0649c1cc7c48de0e68)
